+++ Packages/com.unity.render-pipelines.core/Runtime/Textures/RTHandles.cs
@@ -91,6 +91,81 @@ namespace UnityEngine.Rendering
         }
 
         /// <summary>
+        /// Allocate a new fixed sized RTHandle with the default RTHandle System.
+        /// </summary>
+        /// <param name="width">With of the RTHandle.</param>
+        /// <param name="height">Heigh of the RTHandle.</param>
+        /// <param name="wrapModeU">U coordinate wrapping mode of the RTHandle.</param>
+        /// <param name="wrapModeV">V coordinate wrapping mode of the RTHandle.</param>
+        /// <param name="wrapModeW">W coordinate wrapping mode of the RTHandle.</param>
+        /// <param name="slices">Number of slices of the RTHandle.</param>
+        /// <param name="depthBufferBits">Bit depths of a depth buffer.</param>
+        /// <param name="colorFormat">GraphicsFormat of a color buffer.</param>
+        /// <param name="filterMode">Filtering mode of the RTHandle.</param>
+        /// <param name="dimension">Texture dimension of the RTHandle.</param>
+        /// <param name="enableRandomWrite">Set to true to enable UAV random read writes on the texture.</param>
+        /// <param name="useMipMap">Set to true if the texture should have mipmaps.</param>
+        /// <param name="autoGenerateMips">Set to true to automatically generate mipmaps.</param>
+        /// <param name="isShadowMap">Set to true if the depth buffer should be used as a shadow map.</param>
+        /// <param name="anisoLevel">Anisotropic filtering level.</param>
+        /// <param name="mipMapBias">Bias applied to mipmaps during filtering.</param>
+        /// <param name="msaaSamples">Number of MSAA samples for the RTHandle.</param>
+        /// <param name="bindTextureMS">Set to true if the texture needs to be bound as a multisampled texture in the shader.</param>
+        /// <param name="useDynamicScale">Set to true to use hardware dynamic scaling.</param>
+        /// <param name="memoryless">Use this property to set the render texture memoryless modes.</param>
+        /// <param name="vrUsage">Special treatment of the VR eye texture used in stereoscopic rendering.</param>
+        /// <param name="name">Name of the RTHandle.</param>
+        /// <returns></returns>
+        public static RTHandle Alloc(
+            int width,
+            int height,
+            TextureWrapMode wrapModeU,
+            TextureWrapMode wrapModeV,
+            TextureWrapMode wrapModeW = TextureWrapMode.Repeat,
+            int slices = 1,
+            DepthBits depthBufferBits = DepthBits.None,
+            GraphicsFormat colorFormat = GraphicsFormat.R8G8B8A8_SRGB,
+            FilterMode filterMode = FilterMode.Point,
+            TextureDimension dimension = TextureDimension.Tex2D,
+            bool enableRandomWrite = false,
+            bool useMipMap = false,
+            bool autoGenerateMips = true,
+            bool isShadowMap = false,
+            int anisoLevel = 1,
+            float mipMapBias = 0,
+            MSAASamples msaaSamples = MSAASamples.None,
+            bool bindTextureMS = false,
+            bool useDynamicScale = false,
+            RenderTextureMemoryless memoryless = RenderTextureMemoryless.None,
+            string name = ""
+        )
+        {
+            return s_DefaultInstance.Alloc(
+                width,
+                height,
+                wrapModeU,
+                wrapModeV,
+                wrapModeW,
+                slices,
+                depthBufferBits,
+                colorFormat,
+                filterMode,
+                dimension,
+                enableRandomWrite,
+                useMipMap,
+                autoGenerateMips,
+                isShadowMap,
+                anisoLevel,
+                mipMapBias,
+                msaaSamples,
+                bindTextureMS,
+                useDynamicScale,
+                memoryless,
+                name
+            );
+        }
+
+        /// <summary>
         /// Allocate a new automatically sized RTHandle for the default RTHandle System.
         /// </summary>
         /// <param name="scaleFactor">Constant scale for the RTHandle size computation.</param>