+++ Packages/com.unity.render-pipelines.core/Runtime/Volume/VolumeComponent.cs
@@ -306,6 +306,9 @@ namespace UnityEngine.Rendering
         /// </summary>
         public void Release()
         {
+            if (parameters == null)
+                return;
+
             for (int i = 0; i < parameters.Count; i++)
             {
                 if (parameters[i] != null)