+++ Packages/com.unity.shadergraph/Editor/Generation/Targets/BuiltIn/Editor/ShaderPreprocessor.cs
@@ -111,7 +111,9 @@ namespace UnityEditor.Rendering.BuiltIn
 
         bool IsShaderGraphShader(Shader shader, ShaderSnippetData snippetData)
         {
-            var shaderGraphTag = shader.FindSubshaderTagValue((int)snippetData.pass.SubshaderIndex, m_ShaderGraphShaderTag);
+            var shaderData = ShaderUtil.GetShaderData(shader);
+            var serializedSubShader = shaderData.GetSerializedSubshader((int)snippetData.pass.SubshaderIndex);
+            var shaderGraphTag = serializedSubShader.FindTagValue(m_ShaderGraphShaderTag);
             if (shaderGraphTag == ShaderTagId.none)
                 return false;
             return true;